% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_epoch.R
\name{resample_epoch}
\alias{resample_epoch}
\title{Resample time series data to user-defined epochs}
\usage{
resample_epoch(df, FUN = sum, desired_epoch_s = 60)
}
\arguments{
\item{df}{A data frame with exactly two columns:
- First column: POSIXct timestamps.
- Second column: numeric values to aggregate.}

\item{FUN}{A function to aggregate the numeric values (default is \code{sum}).}

\item{desired_epoch_s}{Length of epoch in seconds (default: 60).}
}
\value{
A data frame with:
  - \code{timestamp_epoch}: the start of each epoch.
  - \code{n_samples}: number of non-NA values in that epoch.
  - \code{value}: aggregated value for the epoch.
}
\description{
Aggregates a numeric variable in a two-column data frame into regular epochs
of user-defined duration (in seconds), returning both the aggregated value
and number of valid samples per epoch.
}
\examples{
df <- data.frame(
  timestamp = as.POSIXct("2025-05-30 18:00:00") + seq(0, 180, by = 5),
  steps = sample(0:5, 37, replace = TRUE)
)
resample_epoch(df, FUN = sum, desired_epoch_s = 60)

}
