% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepinWearable-package.R
\docType{package}
\name{stepinWearable-package}
\alias{stepinWearable}
\alias{stepinWearable-package}
\title{stepinWearable: Personalized Wearable Platform for STEP-IN Project}
\description{
A Shiny application to ingest, process, visualize, and generate
personalized targets from minute-level wearable data (Fitbit, Matrix, StepWatch).
}
\section{Modules}{

- **Intervention**: data loading, t-index, plots, decision engine.
- **Assessment**: (in development).
}

\section{Package options}{

\describe{
  \item{\code{stepin.server_host}}{VPN server (char).}
  \item{\code{stepin.path_windows}}{Windows default path.}
  \item{\code{stepin.path_unix}}{MacOS default path.}
  \item{\code{stepin.output_dir}}{Default output folder.}
  \item{\code{stepin.fitbit_download}}{Default fitbit download folder.}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PROFITH/stepinWearable}
  \item Report bugs at \url{https://github.com/PROFITH/stepinWearable/issues}
}

}
\author{
\strong{Maintainer}: Jairo H. Migueles \email{jairo@jhmigueles.com} (PROFITH Research Group)

Other contributors:
\itemize{
  \item Emilio J. Barranco-Moreno \email{emilio.barranco@ugr.es} (PROFITH Research Group) [contributor]
}

}
\keyword{internal}
